@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { FormCategoryComponent } from './form-category.component';
 
@@ -6,7 +6,7 @@ describe('FormCategoryComponent', () => {
   let component: FormCategoryComponent;
   let fixture: ComponentFixture<FormCategoryComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ FormCategoryComponent ]
     })
