@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { FormSubcategoryComponent } from './form-subcategory.component';
 
@@ -6,7 +6,7 @@ describe('FormSubcategoryComponent', () => {
   let component: FormSubcategoryComponent;
   let fixture: ComponentFixture<FormSubcategoryComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ FormSubcategoryComponent ]
     })
