@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { UploadFormComponent } from './upload-form.component';
 
@@ -6,7 +6,7 @@ describe('UploadFormComponent', () => {
   let component: UploadFormComponent;
   let fixture: ComponentFixture<UploadFormComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ UploadFormComponent ]
     })
